clean-version/requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.6
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
chromadb>=0.4.18

# Document processing
pypdf>=3.17.0
python-docx>=0.8.11
python-pptx>=0.6.21
openpyxl>=3.1.2
pandas>=2.0.0
PyMuPDF>=1.23.0

# Web search and scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
duckduckgo-search>=3.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
tiktoken>=0.5.0
tenacity>=8.2.0
aiohttp>=3.8.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.4.0

# Logging and monitoring
structlog>=23.0.0
rich>=13.0.0

# Optional: for advanced evaluation
deepeval>=0.20.0
